+++ duckscript_sdk/src/lib_test.rs
@@ -1,5 +1,6 @@
 use super::*;
-use duckscript::types::command::Commands;
+use duckscript::runner;
+use duckscript::types::runtime::Context;
 
 #[test]
 fn version_test() {
@@ -10,13 +11,17 @@ fn version_test() {
 
 #[test]
 fn load_valid() {
-    let mut commands = Commands::new();
-    let result = load(&mut commands);
+    let mut context = Context::new();
+    let result = load(&mut context.commands);
 
     match result {
         Ok(_) => (),
         Err(e) => panic!("{:#?}", e),
     };
     //  assert!(result.is_ok());
-    assert!(!commands.get_all_command_names().is_empty());
+    assert!(!context.commands.get_all_command_names().is_empty());
+
+    let result = runner::run_script("test_directory ../test", context);
+
+    assert!(result.is_ok());
 }