+++ duckscript_sdk/src/sdk/std/string/mod.rs
@@ -1,3 +1,4 @@
+mod bytes_to_string;
 mod concat;
 mod contains;
 mod ends_with;
@@ -9,6 +10,7 @@ mod length;
 mod replace;
 mod split;
 mod starts_with;
+pub(crate) mod string_to_bytes;
 mod substring;
 mod trim;
 mod trim_end;
@@ -23,6 +25,7 @@ static PACKAGE: &str = "string";
 pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptError> {
     let package = pckg::concat(parent, PACKAGE);
 
+    commands.set(bytes_to_string::create(&package))?;
     commands.set(concat::create(&package)?)?;
     commands.set(contains::create(&package))?;
     commands.set(ends_with::create(&package))?;
@@ -34,6 +37,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(replace::create(&package))?;
     commands.set(split::create(&package))?;
     commands.set(starts_with::create(&package))?;
+    commands.set(string_to_bytes::create(&package))?;
     commands.set(substring::create(&package))?;
     commands.set(trim::create(&package))?;
     commands.set(trim_start::create(&package))?;